test: ScreeningMovieActivityTest
